address pr comments. Speed up spacecraft pset test
